enhancement: update app layout with system monitor open
